Projekty/Task_Manager_API/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic==2.9.0
pytest==8.3.0
httpx==0.27.0
